# Cache sorted flags