As you can't delete flags in FlagSet, this method works

Gainful performance improvement:
```
benchmark                      old ns/op     new ns/op     delta
BenchmarkVisitAll10Flags-4     1549          33.9          -97.81%
BenchmarkVisit10Flags-4        1704          34.2          -97.99%

benchmark                      old allocs     new allocs     delta
BenchmarkVisitAll10Flags-4     3              0              -100.00%
BenchmarkVisit10Flags-4        3              0              -100.00%

benchmark                      old bytes     new bytes     delta
BenchmarkVisitAll10Flags-4     272           0             -100.00%
BenchmarkVisit10Flags-4        272           0             -100.00%
```