@eparis 2017-04-05
Is this an optimization we should care about? We're making the code (especially the already difficult SetNormalizeFunc) more difficult to maintain. Is there a reason one would really call Visit/VisitAll enough times to make this worth it? I agree and understand that it is faster, but is it really important enough to make the code harder to maintain?