@bogem 2017-04-05
It's very important for cobra, where we should really call `VisitAll ` a lot of times. Just benchmark for a small commands (parent command has only 2 persistent flags):
```
benchmark                     old ns/op     new ns/op     delta
BenchmarkInheritedFlags-4     7055          3203          -54.60%
BenchmarkLocalFlags-4         3598          1607          -55.34%

benchmark                     old allocs     new allocs     delta
BenchmarkInheritedFlags-4     49             16             -67.35%
BenchmarkLocalFlags-4         23             8              -65.22%

benchmark                     old bytes     new bytes     delta
BenchmarkInheritedFlags-4     2040          1392          -31.76%
BenchmarkLocalFlags-4         1008          696           -30.95%
```
You can imagine how huge gain will be for commands with a bunch of persistent/local flags (e.g for hugo or docker)